# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Cal Booking API
  version: 0.0.0
...
paths:
  /api/admin/bookings:
    get:
      operationId: listAdminBookings
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - upcoming
              - past
              - cancelled
              - all
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
                required:
                  - bookings
  /api/admin/event-types:
    get:
      operationId: listAdminEventTypes
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventTypesResponse'
    post:
      operationId: createAdminEventType
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedEventTypeResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventTypeRequest'
...
components:
  schemas:
...
    Booking:
      type: object
      required:
        - id
        - eventTypeId
        - eventType
        - guestName
        - guestEmail
        - startAt
        - endAt
        - status
      properties:
        id:
          type: string
        eventTypeId:
          type: string
        eventType:
          $ref: '#/components/schemas/EventType'
        guestName:
          type: string
        guestEmail:
          type: string
        guestNotes:
          type: string
          nullable: true
        startAt:
          type: string
        endAt:
          type: string
        status:
          $ref: '#/components/schemas/BookingStatus'
        createdAt:
          type: string
        cancelledAt:
          type: string
          nullable: true
...
    BookingStatus:
      type: string
      enum:
        - confirmed
        - cancelled
...
    CreateEventTypeRequest:
      type: object
      required:
        - id
        - title
        - description
        - durationMinutes
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
    CreatedEventTypeResponse:
      type: object
      required:
        - eventType
      properties:
        eventType:
          $ref: '#/components/schemas/EventType'
    EventType:
      type: object
      required:
        - id
        - title
        - description
        - durationMinutes
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
    EventTypesResponse:
      type: object
      required:
        - eventTypes
      properties:
        eventTypes:
          type: array
          items:
            $ref: '#/components/schemas/EventType'
...
servers:
  - url: http://127.0.0.1:3000
    description: Local development server
    variables: {}
  - url: https://your-domain.example
    description: Production server
    variables: {}